[CHORE] Split logger classes to avoid cyclic dependencies (#15559)
